// src/services/types/resources/about_us.ts
export type AboutUsApiResponse = {
  _id: string
  about_us: {
    section1: {
      title1_en: string
      title1_fa: string
      title2_en: string
      title2_fa: string
      text1_en: string
      text1_fa: string
      button1_en: string
      button1_fa: string
    }
    section2: {
      title1_en: string
      title1_fa: string
      text1_en: string
      text1_fa: string
      text2_en: string
      text2_fa: string
    }
    section3: {
      title1_en: string
      title1_fa: string
      title2_en: string
      title2_fa: string
      title3_en: string
      title3_fa: string
      text1_en: string
      text1_fa: string
      text2_en: string
      text2_fa: string
      text3_en: string
      text3_fa: string
      image1: string
      image2: string
      image3: string
    }
    section4: {
      title1_en: string
      title1_fa: string
      image1: string
      image2: string
      image3: string
      image4: string
      image5: string
      image6: string
    }
    section5: {
      title1_en: string
      title1_fa: string
      title2_en: string
      title2_fa: string
      title3_en: string
      title3_fa: string
      title4_en: string
      title4_fa: string
      title5_en: string
      title5_fa: string
      svg1: string
      svg2: string
      svg3: string
      svg4: string
      text1_en: string
      text1_fa: string
      text2_en: string
      text2_fa: string
      text3_en: string
      text3_fa: string
      text4_en: string
      text4_fa: string
    }
    section6: {
      text1_en: string
      text1_fa: string
      text2_en: string
      text2_fa: string
      button1_en: string
      button1_fa: string
    }
  }
  __v: number
}